% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AccountDetailsResponse}
\alias{AccountDetailsResponse}
\title{AccountDetailsResponse}
\usage{
AccountDetailsResponse(currencyCode = NULL, firstName = NULL,
  lastName = NULL, localeCode = NULL, region = NULL, timezone = NULL,
  discountRate = NULL, pointsBalance = NULL)
}
\arguments{
\item{currencyCode}{is optional of type character.
Default user currency Code.}

\item{firstName}{is optional of type character.
First Name.}

\item{lastName}{is optional of type character.
Last Name.}

\item{localeCode}{is optional of type character.
Locale Code.}

\item{region}{is optional of type character.
Region.}

\item{timezone}{is optional of type character.
User Time Zone.}

\item{discountRate}{is optional of type numeric.
User Discount Rate.}

\item{pointsBalance}{is optional of type integer.
The Betfair points balance.}
}
\description{
Response for Account details.
}

