% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AccountFundsResponse}
\alias{AccountFundsResponse}
\title{AccountFundsResponse}
\usage{
AccountFundsResponse(availableToBetBalance = NULL, exposure = NULL,
  retainedCommission = NULL, exposureLimit = NULL, discountRate = NULL,
  pointsBalance = NULL)
}
\arguments{
\item{availableToBetBalance}{is optional of type numeric.
Amount available to bet.}

\item{exposure}{is optional of type numeric.
Current exposure.}

\item{retainedCommission}{is optional of type numeric.
Sum of retained commission.}

\item{exposureLimit}{is optional of type numeric.
Exposure limit.}

\item{discountRate}{is optional of type numeric.
User Discount Rate.}

\item{pointsBalance}{is optional of type integer.
The Betfair points balance.}
}
\description{
Response for retrieving available to bet.
}

