% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AccountStatementReport}
\alias{AccountStatementReport}
\title{AccountStatementReport}
\usage{
AccountStatementReport(accountStatement, moreAvailable)
}
\arguments{
\item{accountStatement}{is mandatory of type ListOfStatementItem.
The list of statement items returned by your request.}

\item{moreAvailable}{is mandatory of type logical.
Indicates whether there are further result items beyond this page.}
}
\description{
A container representing search results.
}

