% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AccountSubscription}
\alias{AccountSubscription}
\title{AccountSubscription}
\usage{
AccountSubscription(subscriptionTokens, applicationName = NULL,
  applicationVersionId = NULL)
}
\arguments{
\item{subscriptionTokens}{is mandatory of type ListOfSubscriptionTokenInfo.
Lis t of subscription token details}

\item{applicationName}{is optional of type character.
Application name}

\item{applicationVersionId}{is optional of type character.
Application version Id}
}
\description{
Application subscription details
}

