% Generated by roxygen2 (4.0.2): do not edit by hand
\name{activateApplicationSubscription}
\alias{activateApplicationSubscription}
\title{activateApplicationSubscription}
\usage{
activateApplicationSubscription(subscriptionToken = string(),
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{subscriptionToken}{is mandatory of type character.
Subscription token for activation}
}
\value{
\code{\link{Status}}
}
\description{
Activate the subscription token for the customer
}

