% Generated by roxygen2 (4.0.2): do not edit by hand
\name{APINGException}
\alias{APINGException}
\title{APINGException}
\usage{
APINGException(errorCode,errorDetails,requestUUID)
}
\arguments{
\item{errorCode}{of type string
the unique code for this error}

\item{errorDetails}{of type string
Specific error details}

\item{requestUUID}{of type string}
}
\description{
This exception is thrown when an operation fails
}
\details{
{
  \itemize{
    \item {     INVALID_INPUT_DATA
           Invalid input data
    }
    \item {     INVALID_SESSION_INFORMATION
           The session token passed is invalid
    }
    \item {      NO_APP_KEY
           An application key is required for this operation
    }
    \item {     NO_SESSION
           A session token is required for this operation
    }
    \item {      INVALID_APP_KEY
           The application key passed is invalid
    }
    \item {     UNEXPECTED_ERROR
           An unexpected internal error occurred that prevented successful
           request processing.
    }
  }
}
}

