% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{BetStatusValidValues}
\alias{BetStatusValidValues}
\title{BetStatus}
\format{\preformatted{ chr [1:4] "SETTLED" "VOIDED" "LAPSED" "CANCELLED"
}}
\usage{
BetStatus(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
BetStatus
}
\details{
{
  \itemize{
    \item {     SETTLED
           A matched bet that was settled normally
    }
    \item {     VOIDED
           A matched bet that was subsequently voided by Betfair, before, during
           or after settlement
    }
    \item {     LAPSED
           Unmatched bet that was cancelled by Betfair (for example at turn in
           play).
    }
    \item {     CANCELLED
           Unmatched bet that was cancelled by an explicit customer action.
    }
  }
}
}
\keyword{datasets}

