% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cancelApplicationSubscription}
\alias{cancelApplicationSubscription}
\title{cancelApplicationSubscription}
\usage{
cancelApplicationSubscription(subscriptionToken = string(),
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{subscriptionToken}{is mandatory of type character.
Subscription token to cancel}
}
\value{
\code{\link{Status}}
}
\description{
Cancel application subscription
}

