% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CancelInstruction}
\alias{CancelInstruction}
\title{CancelInstruction}
\usage{
CancelInstruction(betId, sizeReduction = NULL)
}
\arguments{
\item{betId}{is mandatory of type character.}

\item{sizeReduction}{is optional of type numeric.
If supplied then this is a partial cancel}
}
\description{
Instruction to fully or partially cancel an order (only applies to LIMIT
orders)
}

