% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CancelInstructionReport}
\alias{CancelInstructionReport}
\title{CancelInstructionReport}
\usage{
CancelInstructionReport(status, errorCode = NULL, instruction = NULL,
  sizeCancelled, cancelledDate)
}
\arguments{
\item{status}{is mandatory of type character.
whether the command succeeded or failed}

\item{errorCode}{is optional of type character.
cause of failure, or null if command succeeds}

\item{instruction}{is optional of type CancelInstruction.
The instruction that was requested}

\item{sizeCancelled}{is mandatory of type numeric.}

\item{cancelledDate}{is mandatory of type dateTime.}
}
\description{
CancelInstructionReport
}

