% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ClearedOrderSummary}
\alias{ClearedOrderSummary}
\title{ClearedOrderSummary}
\usage{
ClearedOrderSummary(eventTypeId = NULL, eventId = NULL, marketId = NULL,
  selectionId = NULL, handicap = NULL, betId = NULL, placedDate = NULL,
  persistenceType = NULL, orderType = NULL, side = NULL,
  itemDescription = NULL, priceRequested = NULL, settledDate = NULL,
  betCount = NULL, commission = NULL, priceMatched = NULL,
  priceReduced = NULL, sizeSettled = NULL, profit = NULL,
  sizeCancelled = NULL)
}
\arguments{
\item{eventTypeId}{is optional of type character.
The id of the event type bet on. Available at EVENT_TYPE groupBy level
or lower.}

\item{eventId}{is optional of type character.
The id of the event bet on. Available at EVENT groupBy level or lower.}

\item{marketId}{is optional of type character.
The id of the market bet on. Available at MARKET groupBy level or
lower.}

\item{selectionId}{is optional of type integer.
The id of the selection bet on. Available at RUNNER groupBy level or
lower.}

\item{handicap}{is optional of type numeric.
The id of the market bet on. Available at MARKET groupBy level or
lower.}

\item{betId}{is optional of type character.
The id of the bet. Available at BET groupBy level.}

\item{placedDate}{is optional of type dateTime.
The date the bet order was placed by the customer. Only available at
BET groupBy level.}

\item{persistenceType}{is optional of type character.
The turn in play persistence state of the order at bet placement time.
This field will be empty or omitted on true SP bets. Only available at
BET groupBy level.}

\item{orderType}{is optional of type character.
The type of bet (e.g standard limited-liability Exchange bet (LIMIT),
a standard BSP bet (MARKET_ON_CLOSE), or a minimum-accepted-price BSP
bet (LIMIT_ON_CLOSE)). If the bet has a OrderType of MARKET_ON_CLOSE
and a persistenceType of MARKET_ON_CLOSE then it is a bet which has
transitioned from LIMIT to MARKET_ON_CLOSE. Only available at BET
groupBy level.}

\item{side}{is optional of type character.
Whether the bet was a back or lay bet. Available at SIDE groupBy level
or lower.}

\item{itemDescription}{is optional of type ItemDescription.
A container for all the anciliary data and localised text valid for
this Item}

\item{priceRequested}{is optional of type numeric.
The average requested price across all settled bet orders under this
Item. Available at SIDE groupBy level or lower.}

\item{settledDate}{is optional of type dateTime.
The date and time the bet order was settled by Betfair. Available at
SIDE groupBy level or lower.}

\item{betCount}{is optional of type integer.
The number of actual bets within this grouping (will be 1 for BET
groupBy)}

\item{commission}{is optional of type numeric.
The cumulative amount of commission paid by the customer across all
bets under this Item, in the account currency. Available at EXCHANGE,
EVENT_TYPE, EVENT and MARKET level groupings only.}

\item{priceMatched}{is optional of type numeric.
The average matched price across all settled bets or bet fragments
under this Item. Available at SIDE groupBy level or lower.}

\item{priceReduced}{is optional of type logical.
If true, then the matched price was affected by a reduction factor due
to of a runner removal from this Horse Racing market.}

\item{sizeSettled}{is optional of type numeric.
The cumulative bet size that was settled as matched or voided under
this Item, in the account currency. Available at SIDE groupBy level or
lower.}

\item{profit}{is optional of type numeric.
The profit or loss (negative profit) gained on this line, in the
account currency}

\item{sizeCancelled}{is optional of type numeric.
The amount of the bet that was available to be matched, before
cancellation or lapsing, in the account currency}
}
\description{
Summary of a cleared order.
}

