% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CompetitionResult}
\alias{CompetitionResult}
\title{CompetitionResult}
\usage{
CompetitionResult(competition = NULL, marketCount = NULL,
  competitionRegion = NULL)
}
\arguments{
\item{competition}{is optional of type Competition.
Competition}

\item{marketCount}{is optional of type integer.
Count of markets associated with this competition}

\item{competitionRegion}{is optional of type character.
Region in which this competition is happening}
}
\description{
Competition Result
}

