% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CountryCodeResult}
\alias{CountryCodeResult}
\title{CountryCodeResult}
\usage{
CountryCodeResult(countryCode = NULL, marketCount = NULL)
}
\arguments{
\item{countryCode}{is optional of type character.
Country Code}

\item{marketCount}{is optional of type integer.
Count of markets associated with this Country Code}
}
\description{
CountryCode Result
}

