% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createDeveloperAppKeys}
\alias{createDeveloperAppKeys}
\title{createDeveloperAppKeys}
\usage{
createDeveloperAppKeys(appName = string(), exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{appName}{is mandatory of type character.
A Display name for the application.}
}
\value{
A map of application keys, one marked ACTIVE, and the other DELAYED
\code{\link{DeveloperApp}}
}
\description{
Create 2 application keys for given user; one active and the other delayed
}

