% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CurrencyRate}
\alias{CurrencyRate}
\title{CurrencyRate}
\usage{
CurrencyRate(currencyCode = NULL, rate = NULL)
}
\arguments{
\item{currencyCode}{is optional of type character.
Three letter ISO 4217 code}

\item{rate}{is optional of type numeric.
Exchange rate for the currency specified in the request}
}
\description{
Currency rate
}

