% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CurrentOrderSummary}
\alias{CurrentOrderSummary}
\title{CurrentOrderSummary}
\usage{
CurrentOrderSummary(betId, marketId, selectionId, handicap, priceSize,
  bspLiability, side, status, persistenceType, orderType, placedDate,
  matchedDate, averagePriceMatched = NULL, sizeMatched = NULL,
  sizeRemaining = NULL, sizeLapsed = NULL, sizeCancelled = NULL,
  sizeVoided = NULL, regulatorAuthCode = NULL, regulatorCode = NULL)
}
\arguments{
\item{betId}{is mandatory of type character.
The bet ID of the original place order.}

\item{marketId}{is mandatory of type character.
The market id the order is for.}

\item{selectionId}{is mandatory of type integer.
The selection id the order is for.}

\item{handicap}{is mandatory of type numeric.
The handicap of the bet.}

\item{priceSize}{is mandatory of type PriceSize.
The price and size of the bet.}

\item{bspLiability}{is mandatory of type numeric.
Not to be confused with size. This is the liability of a given BSP
bet.}

\item{side}{is mandatory of type character.
BACK/LAY}

\item{status}{is mandatory of type character.
Either EXECUTABLE (an unmatched amount remains) or EXECUTION_COMPLETE
(no unmatched amount remains).}

\item{persistenceType}{is mandatory of type character.
What to do with the order at turn-in-play.}

\item{orderType}{is mandatory of type character.
BSP Order type.}

\item{placedDate}{is mandatory of type dateTime.
The date, to the second, the bet was placed.}

\item{matchedDate}{is mandatory of type dateTime.
The date, to the second, of the last matched bet fragment (where
applicable).}

\item{averagePriceMatched}{is optional of type numeric.
The average price matched at. Voided match fragments are removed from
this average calculation.}

\item{sizeMatched}{is optional of type numeric.
The current amount of this bet that was matched.}

\item{sizeRemaining}{is optional of type numeric.
The current amount of this bet that is unmatched.}

\item{sizeLapsed}{is optional of type numeric.
The current amount of this bet that was lapsed.}

\item{sizeCancelled}{is optional of type numeric.
The current amount of this bet that was cancelled.}

\item{sizeVoided}{is optional of type numeric.
The current amount of this bet that was voided.}

\item{regulatorAuthCode}{is optional of type character.
The regulator authorisation code.}

\item{regulatorCode}{is optional of type character.
The regulator Code.}
}
\description{
Summary of a current order.
}

