% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CurrentOrderSummaryReport}
\alias{CurrentOrderSummaryReport}
\title{CurrentOrderSummaryReport}
\usage{
CurrentOrderSummaryReport(currentOrders, moreAvailable)
}
\arguments{
\item{currentOrders}{is mandatory of type ListOfCurrentOrderSummary.
The list of current orders returned by your query. This will be a
valid list (i.e. empty or non-empty but never 'null').}

\item{moreAvailable}{is mandatory of type logical.
Indicates whether there are further result items beyond this page.
Note that underlying data is highly time-dependent and the subsequent
search orders query might return an empty result.}
}
\description{
A container representing search results.
}

