% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DeveloperApp}
\alias{DeveloperApp}
\title{DeveloperApp}
\usage{
DeveloperApp(appName, appId, appVersions)
}
\arguments{
\item{appName}{is mandatory of type character.
The unique name of the application}

\item{appId}{is mandatory of type integer.
A unique id of this application}

\item{appVersions}{is mandatory of type ListOfDeveloperAppVersion.
The application versions (including application keys)}
}
\description{
Describes developer/vendor specific application
}

