% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DeveloperAppVersion}
\alias{DeveloperAppVersion}
\title{DeveloperAppVersion}
\usage{
DeveloperAppVersion(owner, versionId, version, applicationKey,
  delayData = NULL, subscriptionRequired, ownerManaged, active)
}
\arguments{
\item{owner}{is mandatory of type character.
The sportex user who owns the specific version of the application}

\item{versionId}{is mandatory of type integer.
The unique Id of the application version}

\item{version}{is mandatory of type character.
The version identifier string such as 1.0, 2.0. Unique for a given
application.}

\item{applicationKey}{is mandatory of type character.
The unqiue application key associated with this application version}

\item{delayData}{is optional of type logical.
Indicates whether the data exposed by platform services as seen by
this application key is delayed or realtime.}

\item{subscriptionRequired}{is mandatory of type logical.
Indicates whether the application version needs explicit subscription}

\item{ownerManaged}{is mandatory of type logical.
Indicates whether the application version needs explicit management by
producers. A value of false indicates, this is a version meant for
developer use.}

\item{active}{is mandatory of type logical.
Indicates whether the application version is currently active}
}
\description{
Describes a version of an external application
}

