% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Event}
\alias{Event}
\title{Event}
\usage{
Event(id = NULL, name = NULL, countryCode = NULL, timezone = NULL,
  venue = NULL, openDate = NULL)
}
\arguments{
\item{id}{is optional of type character.
The unique id for the event}

\item{name}{is optional of type character.
The name of the event}

\item{countryCode}{is optional of type character.
countryCode}

\item{timezone}{is optional of type character.
timezone}

\item{venue}{is optional of type character.
venue}

\item{openDate}{is optional of type dateTime.
openDate}
}
\description{
Event
}

