% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EventResult}
\alias{EventResult}
\title{EventResult}
\usage{
EventResult(event = NULL, marketCount = NULL)
}
\arguments{
\item{event}{is optional of type Event.
Event}

\item{marketCount}{is optional of type integer.
Count of markets associated with this event}
}
\description{
Event Result
}

