% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EventTypeResult}
\alias{EventTypeResult}
\title{EventTypeResult}
\usage{
EventTypeResult(eventType = NULL, marketCount = NULL)
}
\arguments{
\item{eventType}{is optional of type EventType.
EventType}

\item{marketCount}{is optional of type integer.
Count of markets associated with this eventType}
}
\description{
EventType Result
}

