% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExBestOffersOverrides}
\alias{ExBestOffersOverrides}
\title{ExBestOffersOverrides}
\usage{
ExBestOffersOverrides(bestPricesDepth = NULL, rollupModel = NULL,
  rollupLimit = NULL, rollupLiabilityThreshold = NULL,
  rollupLiabilityFactor = NULL)
}
\arguments{
\item{bestPricesDepth}{is optional of type integer.
The maximum number of prices to return on each side for each runner.
If unspecified defaults to 3.}

\item{rollupModel}{is optional of type RollupModel.
The model to use when rolling up available sizes. If unspecified
defaults to STAKE rollup model with rollupLimit of minimum stake in
the specified currency.}

\item{rollupLimit}{is optional of type integer.
The volume limit to use when rolling up returned sizes. The exact
definition of the limit depends on the rollupModel. If no limit is
provided it will use minimum stake as default the value. Ignored if no
rollup model is specified.}

\item{rollupLiabilityThreshold}{is optional of type numeric.
Only applicable when rollupModel is MANAGED_LIABILITY. The rollup
model switches from being stake based to liability based at the
smallest lay price which is >= rollupLiabilityThreshold.service level
default (TBD). Not supported as yet.}

\item{rollupLiabilityFactor}{is optional of type integer.
Only applicable when rollupModel is MANAGED_LIABILITY.
(rollupLiabilityFactor * rollupLimit) is the minimum liabilty the user
is deemed to be comfortable with. After the rollupLiabilityThreshold
price subsequent volumes will be rolled up to minimum value such that
the liability >= the minimum liability.service level default (5). Not
supported as yet.}
}
\description{
Options to alter the default representation of best offer prices
}

