% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExchangePrices}
\alias{ExchangePrices}
\title{ExchangePrices}
\usage{
ExchangePrices(availableToBack = NULL, availableToLay = NULL,
  tradedVolume = NULL)
}
\arguments{
\item{availableToBack}{is optional of type ListOfPriceSize.}

\item{availableToLay}{is optional of type ListOfPriceSize.}

\item{tradedVolume}{is optional of type ListOfPriceSize.}
}
\description{
ExchangePrices
}

