% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ExecutionReportErrorCodeValidValues}
\alias{ExecutionReportErrorCodeValidValues}
\title{ExecutionReportErrorCode}
\format{\preformatted{ chr [1:17] "ERROR_IN_MATCHER" "PROCESSED_WITH_ERRORS" "BET_ACTION_ERROR" "INVALID_ACCOUNT_STATE" ...
}}
\usage{
ExecutionReportErrorCode(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
ExecutionReportErrorCode
}
\details{
{
  \itemize{
    \item {     ERROR_IN_MATCHER
           The matcher's not healthy
    }
    \item {     PROCESSED_WITH_ERRORS
           The order itself has been accepted, but at least one (possibly all)
           actions have generated errors
    }
    \item {     BET_ACTION_ERROR
           There is an error with an action that has caused the entire order to
           be rejected
    }
    \item {     INVALID_ACCOUNT_STATE
           Order rejected due to the account's status (suspended, inactive, dup
           cards)
    }
    \item {     INVALID_WALLET_STATUS
           Order rejected due to the account's wallet's status
    }
    \item {     INSUFFICIENT_FUNDS
           Account has exceeded its exposure limit or available to bet limit
    }
    \item {     LOSS_LIMIT_EXCEEDED
           The account has exceed the self imposed loss limit
    }
    \item {     MARKET_SUSPENDED
           Market is suspended
    }
    \item {     MARKET_NOT_OPEN_FOR_BETTING
           Market is not open for betting, either inactive, suspended or closed
    }
    \item {     DUPLICATE_TRANSACTION
           duplicate customer referece data submitted
    }
    \item {     INVALID_ORDER
           Order cannot be accepted by the matcher due to the combination of
           actions. For example, bets being edited are not on the same market, or
           order includes both edits and placement
    }
    \item {     INVALID_MARKET_ID
           Market doesn't exist
    }
    \item {     PERMISSION_DENIED
           Business rules do not allow order to be placed
    }
    \item {     DUPLICATE_BETIDS
           duplicate bet ids found
    }
    \item {     NO_ACTION_REQUIRED
           Order hasn't been passed to matcher as system detected there will be
           no state change
    }
    \item {     SERVICE_UNAVAILABLE
           The requested service is unavailable
    }
    \item {     REJECTED_BY_REGULATOR
           The regulator rejected the order
    }
  }
}
}
\keyword{datasets}

