% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ExecutionReportStatusValidValues}
\alias{ExecutionReportStatusValidValues}
\title{ExecutionReportStatus}
\format{\preformatted{ chr [1:4] "SUCCESS" "FAILURE" "PROCESSED_WITH_ERRORS" "TIMEOUT"
}}
\usage{
ExecutionReportStatus(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
ExecutionReportStatus
}
\details{
{
  \itemize{
    \item {     SUCCESS
    }
    \item {     FAILURE
    }
    \item {     PROCESSED_WITH_ERRORS
    }
    \item {     TIMEOUT
    }
  }
}
}
\keyword{datasets}

