% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAccountDetails}
\alias{getAccountDetails}
\title{getAccountDetails}
\usage{
getAccountDetails(exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\value{
Response for retrieving account details.
\code{\link{AccountDetailsResponse}}
}
\description{
Get Account details.
}

