% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAccountFunds}
\alias{getAccountFunds}
\title{getAccountFunds}
\usage{
getAccountFunds(exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\value{
Response for retrieving available to bet.
\code{\link{AccountFundsResponse}}
}
\description{
Get available to bet amount.
}

