% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAccountStatement}
\alias{getAccountStatement}
\title{getAccountStatement}
\usage{
getAccountStatement(locale = NULL, fromRecord = NULL, recordCount = NULL,
  itemDateRange = NULL, includeItem = NULL, wallet = NULL,
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{locale}{is optional of type character.
The language to be used where applicable. If not specified, the
customer account default is returned.}

\item{fromRecord}{is optional of type integer.
Specifies the first record that will be returned. Records start at
index zero. If not specified then it will default to 0.}

\item{recordCount}{is optional of type integer.
Specifies the maximum number of records to be returned. Note that
there is a page size limit of 100. If not specified then it will
default to the page limit size.}

\item{itemDateRange}{is optional of type TimeRange.
Return items with an itemDate within this date range. Both from and to
date times are inclusive. If from is not specified then the oldest
available items will be in range. If to is not specified then the
latest items will be in range. nb. This itemDataRange is currently
only applied when includeItem is set to ALL or not specified, else
items are NOT bound by itemDate.}

\item{includeItem}{is optional of type IncludeItem.
Which items to include, if not specified then defaults to ALL.}

\item{wallet}{is optional of type Wallet.
Which wallet to return statementItems for. If unspecified then the UK
wallet will be selected}
}
\value{
List of statement items chronologically ordered plus moreAvailable
         boolean to facilitate paging
         \code{\link{AccountStatementReport}}
}
\description{
Get account statement
}

