% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getApplicationSubscriptionToken}
\alias{getApplicationSubscriptionToken}
\title{getApplicationSubscriptionToken}
\usage{
getApplicationSubscriptionToken(subscriptionLength = NULL,
  clientReference = NULL, exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{subscriptionLength}{is optional of type integer.
How many days the subscription should last. Open ended if value not
supplied. Expiry time will be rounded up to midnight on the date of
expiry.}

\item{clientReference}{is optional of type character.
Any client reference for this subscription token request.}
}
\value{
Subscription token
}
\description{
Returns the newly generate subscription token
}

