% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getVendorClientId}
\alias{getVendorClientId}
\title{getVendorClientId}
\usage{
getVendorClientId(exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\value{
Vendor client id.
}
\description{
Get vendor client id for customer account.
}

