% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{GroupByValidValues}
\alias{GroupByValidValues}
\title{GroupBy}
\format{\preformatted{ chr [1:6] "EVENT_TYPE" "EVENT" "MARKET" "RUNNER" "SIDE" "BET"
}}
\usage{
GroupBy(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
GroupBy
}
\details{
{
  \itemize{
    \item {     EVENT_TYPE
           A roll up of settled P&L, commission paid and number of bet orders, on
           a specified event type
    }
    \item {     EVENT
           A roll up of settled P&L, commission paid and number of bet orders, on
           a specified event
    }
    \item {     MARKET
           A roll up of settled P&L, commission paid and number of bet orders, on
           a specified market
    }
    \item {     RUNNER
           A roll up of settled P&L and the number of bet orders, on a specified
           runner within a specified market
    }
    \item {     SIDE
           An averaged roll up of settled P&L, and number of bets, on the
           specified side of a specified selection within a specified market,
           that are either settled or voided
    }
    \item {     BET
           The P&L, commission paid, side and regulatory information etc, about
           each individual bet order
    }
  }
}
}
\keyword{datasets}

