% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HeartbeatReport}
\alias{HeartbeatReport}
\title{HeartbeatReport}
\usage{
HeartbeatReport(actionPerformed, actualTimeoutSeconds)
}
\arguments{
\item{actionPerformed}{is mandatory of type ActionPerformed.
The action performed since your last heartbeat request.}

\item{actualTimeoutSeconds}{is mandatory of type integer.
The actual timeout applied to your heartbeat request, see timeout
request parameter description for details.}
}
\description{
Response from heartbeat operation
}

