% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{IncludeItemValidValues}
\alias{IncludeItemValidValues}
\title{IncludeItem}
\format{\preformatted{ chr [1:4] "ALL" "EXCHANGE" "POKER_ROOM" "DEPOSITS_WITHDRAWALS"
}}
\usage{
IncludeItem(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
IncludeItem
}
\details{
{
  \itemize{
    \item {     ALL
    }
    \item {     EXCHANGE
    }
    \item {     POKER_ROOM
    }
    \item {     DEPOSITS_WITHDRAWALS
    }
  }
}
}
\keyword{datasets}

