% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{InstructionReportErrorCodeValidValues}
\alias{InstructionReportErrorCodeValidValues}
\title{InstructionReportErrorCode}
\format{\preformatted{ chr [1:20] "INVALID_BET_SIZE" "INVALID_RUNNER" "BET_TAKEN_OR_LAPSED" "BET_IN_PROGRESS" "RUNNER_REMOVED" ...
}}
\usage{
InstructionReportErrorCode(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
InstructionReportErrorCode
}
\details{
{
  \itemize{
    \item {     INVALID_BET_SIZE
           Bet size is invalid for your currency or your regulator
    }
    \item {     INVALID_RUNNER
           Runner does not exist, includes vacant traps in greyhound racing
    }
    \item {     BET_TAKEN_OR_LAPSED
           Bet cannot be cancelled or modified as it has already been taken or
           has lapsed Includes attempts to cancel/modify market on close BSP bets
           and cancelling limit on close BSP bets
    }
    \item {     BET_IN_PROGRESS
           No result was received from the matcher in a timeout configured for
           the system
    }
    \item {     RUNNER_REMOVED
           Runner has been removed from the event
    }
    \item {     MARKET_NOT_OPEN_FOR_BETTING
           Attempt to edit a bet on a market that has closed.
    }
    \item {     LOSS_LIMIT_EXCEEDED
           The action has caused the account to exceed the self imposed loss
           limit
    }
    \item {     MARKET_NOT_OPEN_FOR_BSP_BETTING
           Market now closed to bsp betting. Turned in-play or has been
           reconciled
    }
    \item {     INVALID_PRICE_EDIT
           Attempt to edit down the price of a bsp limit on close lay bet, or
           edit up the price of a limit on close back bet
    }
    \item {     INVALID_ODDS
           Odds not on price ladder - either edit or placement
    }
    \item {     INSUFFICIENT_FUNDS
           Insufficient funds available to cover the bet action. Either the
           exposure limit or available to bet limit would be exceeded
    }
    \item {     INVALID_PERSISTENCE_TYPE
           Invalid persistence type for this market, e.g. KEEP for a non bsp
           market
    }
    \item {     ERROR_IN_MATCHER
           A problem with the matcher prevented this action completing
           successfully
    }
    \item {     INVALID_BACK_LAY_COMBINATION
           The order contains a back and a lay for the same runner at overlapping
           prices. This would guarantee a self match. This also applies to BSP
           limit on close bets
    }
    \item {     ERROR_IN_ORDER
           The action failed because the parent order failed
    }
    \item {     INVALID_BID_TYPE
           Bid type is mandatory
    }
    \item {     INVALID_BET_ID
           Bet for id supplied has not been found
    }
    \item {     CANCELLED_NOT_PLACED
           Bet cancelled but replacement bet was not placed
    }
    \item {     RELATED_ACTION_FAILED
           Action failed due to the failure of a action on which this action is
           dependent
    }
    \item {     NO_ACTION_REQUIRED
           the action does not result in any state change. eg changing a
           persistence to it's current value
    }
  }
}
}
\keyword{datasets}

