% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{InstructionReportStatusValidValues}
\alias{InstructionReportStatusValidValues}
\title{InstructionReportStatus}
\format{\preformatted{ chr [1:3] "SUCCESS" "FAILURE" "TIMEOUT"
}}
\usage{
InstructionReportStatus(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
InstructionReportStatus
}
\details{
{
  \itemize{
    \item {     SUCCESS
    }
    \item {     FAILURE
    }
    \item {     TIMEOUT
    }
  }
}
}
\keyword{datasets}

