% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ItemClassValidValues}
\alias{ItemClassValidValues}
\title{ItemClass}
\format{\preformatted{ chr "UNKNOWN"
}}
\usage{
ItemClass(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
ItemClass
}
\details{
{
  \itemize{
    \item {     UNKNOWN
           Statement item not mapped to a specific class. All values will be
           concatenated into a single key/value pair. The key will be
           'unknownStatementItem' and the value will be a comma separated string.

    }
  }
}
}
\keyword{datasets}

