% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ItemDescription}
\alias{ItemDescription}
\title{ItemDescription}
\usage{
ItemDescription(eventTypeDesc = NULL, eventDesc = NULL, marketDesc = NULL,
  marketStartTime = NULL, runnerDesc = NULL, numberOfWinners = NULL)
}
\arguments{
\item{eventTypeDesc}{is optional of type character.
The event type name, translated into the requested locale. Available
at EVENT_TYPE groupBy or lower.}

\item{eventDesc}{is optional of type character.
The eventName, or openDate + venue, translated into the requested
locale. Available at EVENT groupBy or lower.}

\item{marketDesc}{is optional of type character.
The market name or racing market type ("Win", "To Be Placed (2
places)", "To Be Placed (5 places)" etc) translated into the requested
locale. Available at MARKET groupBy or lower.}

\item{marketStartTime}{is optional of type dateTime.
The start time of the market (in ISO-8601 format, not translated).
Available at MARKET groupBy or lower.}

\item{runnerDesc}{is optional of type character.
The runner name, maybe including the handicap, translated into the
requested locale. Available at BET groupBy.}

\item{numberOfWinners}{is optional of type integer.
The numberOfWinners on a market. Available at BET groupBy.}
}
\description{
This object contains some text which may be useful to render a betting
history view. It offers no long-term warranty as to the correctness of the
text.
}

