% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.activateApplicationSubscription}
\alias{json.activateApplicationSubscription}
\title{activateApplicationSubscription}
\usage{
json.activateApplicationSubscription(subscriptionToken = string(),
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{subscriptionToken}{is mandatory of type character.
Subscription token for activation}
}
\description{
Activate the subscription token for the customer
}

