% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.cancelOrders}
\alias{json.cancelOrders}
\title{cancelOrders}
\usage{
json.cancelOrders(marketId = NULL, instructions = NULL,
  customerRef = NULL, exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{marketId}{is optional of type character.
If not supplied all bets are cancelled}

\item{instructions}{is optional of type ListOfCancelInstruction.
All instructions need to be on the same market. If not supplied all
bets on the market (if market id is passed) are fully cancelled}

\item{customerRef}{is optional of type character.
Optional parameter allowing the client to pass a unique string (upto
32 chars) that is used to de-dupe mistaken re-submitions.}
}
\description{
Cancel all bets OR cancel all bets on a market OR fully or partially cancel
particular orders on a market. Only LIMIT orders an be cancelled or partially
cancelled once placed.
}

