% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.createDeveloperAppKeys}
\alias{json.createDeveloperAppKeys}
\title{createDeveloperAppKeys}
\usage{
json.createDeveloperAppKeys(appName = string(),
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{appName}{is mandatory of type character.
A Display name for the application.}
}
\value{
A map of application keys, one marked ACTIVE, and the other DELAYED
}
\description{
Create 2 application keys for given user; one active and the other delayed
}

