% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.getAccountDetails}
\alias{json.getAccountDetails}
\title{getAccountDetails}
\usage{
json.getAccountDetails(exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\value{
Response for retrieving account details.
}
\description{
Get Account details.
}

