% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.getAccountFunds}
\alias{json.getAccountFunds}
\title{getAccountFunds}
\usage{
json.getAccountFunds(exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\value{
Response for retrieving available to bet.
}
\description{
Get available to bet amount.
}

