% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.listAccountSubscriptionTokens}
\alias{json.listAccountSubscriptionTokens}
\title{listAccountSubscriptionTokens}
\usage{
json.listAccountSubscriptionTokens(exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\value{
List of subscription tokens associated with the account
}
\description{
List of subscription tokens associated with the account
}

