% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.listClearedOrders}
\alias{json.listClearedOrders}
\title{listClearedOrders}
\usage{
json.listClearedOrders(betStatus = BetStatus(), eventTypeIds = NULL,
  eventIds = NULL, marketIds = NULL, runnerIds = NULL, betIds = NULL,
  side = NULL, settledDateRange = NULL, groupBy = NULL,
  includeItemDescription = NULL, locale = NULL, fromRecord = NULL,
  recordCount = NULL, exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{betStatus}{is mandatory of type BetStatus.
Restricts the results to the specified status.}

\item{eventTypeIds}{is optional of type SetOfEventTypeId.
Optionally restricts the results to the specified Event Type IDs.}

\item{eventIds}{is optional of type SetOfEventId.
Optionally restricts the results to the specified Event IDs.}

\item{marketIds}{is optional of type SetOfMarketId.
Optionally restricts the results to the specified market IDs.}

\item{runnerIds}{is optional of type SetOfRunnerId.
Optionally restricts the results to the specified Runners.}

\item{betIds}{is optional of type SetOfBetId.
Optionally restricts the results to the specified bet IDs.}

\item{side}{is optional of type Side.
Optionally restricts the results to the specified side.}

\item{settledDateRange}{is optional of type TimeRange.
Optionally restricts the results to be from/to the specified settled
date. This date is inclusive, i.e. if an order was placed on exactly
this date (to the millisecond) then it will be included in the
results. If the from is later than the to, no results will be
returned.}

\item{groupBy}{is optional of type GroupBy.
How to aggregate the lines, if not supplied then the lowest level is
returned, i.e. bet by bet This is only applicable to SETTLED
BetStatus.}

\item{includeItemDescription}{is optional of type logical.
If true then an ItemDescription object is included in the response.}

\item{locale}{is optional of type character.
The language used for the itemDescription. If not specified, the
customer account default is returned.}

\item{fromRecord}{is optional of type integer.
Specifies the first record that will be returned. Records start at
index zero.}

\item{recordCount}{is optional of type integer.
Specifies how many records will be returned, from the index position
'fromRecord'. Note that there is a page size limit of 1000. A value of
zero indicates that you would like all records (including and from
'fromRecord') up to the limit.}
}
\description{
Returns a List of bets based on the bet status, ordered by settled date
}

