% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.listCurrencyRates}
\alias{json.listCurrencyRates}
\title{listCurrencyRates}
\usage{
json.listCurrencyRates(fromCurrency = NULL, exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{fromCurrency}{is optional of type character.
The currency from which the rates are computed. GBP is used if no
value is passed}
}
\value{
List of currency rates
}
\description{
Returns a list of currency rates based on given currency
}

