% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.listMarketProfitAndLoss}
\alias{json.listMarketProfitAndLoss}
\title{listMarketProfitAndLoss}
\usage{
json.listMarketProfitAndLoss(marketIds = SetOfMarketId(),
  includeSettledBets = NULL, includeBspBets = NULL,
  netOfCommission = NULL, exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{marketIds}{is mandatory of type SetOfMarketId.
List of markets to calculate profit and loss}

\item{includeSettledBets}{is optional of type logical.
Option to include settled bets (partially settled markets only).
Defaults to false if not specified.}

\item{includeBspBets}{is optional of type logical.
Option to include BSP bets. Defaults to false if not specified.}

\item{netOfCommission}{is optional of type logical.
Option to return profit and loss net of users current commission rate
for this market including any special tariffs. Defaults to false if
not specified.}
}
\description{
Retrieve profit and loss for a given list of markets. The values are
calculated using matched bets and optionally settled bets. Only odds markets
are implemented, markets of other types are silently ignored.
}

