% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.listTimeRanges}
\alias{json.listTimeRanges}
\title{listTimeRanges}
\usage{
json.listTimeRanges(filter = MarketFilter(),
  granularity = TimeGranularity(), exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{filter}{is mandatory of type MarketFilter.
The filter to select desired markets. All markets that match the
criteria in the filter are selected.}

\item{granularity}{is mandatory of type TimeGranularity.
The granularity of time periods that correspond to markets selected by
the market filter.}
}
\value{
output data
}
\description{
Returns a list of time ranges in the granularity specified in the request
(i.e. 3PM to 4PM, Aug 14th to Aug 15th) associated with the markets selected
by the MarketFilter.
}

