% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.placeOrders}
\alias{json.placeOrders}
\title{placeOrders}
\usage{
json.placeOrders(marketId = MarketId(),
  instructions = ListOfPlaceInstruction(), customerRef = NULL,
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{marketId}{is mandatory of type character.
The market id these orders are to be placed on}

\item{instructions}{is mandatory of type ListOfPlaceInstruction.}

\item{customerRef}{is optional of type character.
Optional parameter allowing the client to pass a unique string (up to
32 chars) that is used to de-dupe mistaken re-submissions.}
}
\description{
Place new orders into market. LIMIT orders below the minimum bet size are
allowed if there is an unmatched bet at the same price in the market. This
operation is atomic in that all orders will be placed or none will be placed.
}

