% Generated by roxygen2 (4.0.2): do not edit by hand
\name{json.replaceOrders}
\alias{json.replaceOrders}
\title{replaceOrders}
\usage{
json.replaceOrders(marketId = MarketId(),
  instructions = ListOfReplaceInstruction(), customerRef = NULL,
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{marketId}{is mandatory of type character.
The market id these orders are to be placed on}

\item{instructions}{is mandatory of type ListOfReplaceInstruction.}

\item{customerRef}{is optional of type character.
Optional parameter allowing the client to pass a unique string (up to
32 chars) that is used to de-dupe mistaken re-submissions.}
}
\description{
This operation is logically a bulk cancel followed by a bulk place. The
cancel is completed first then the new orders are placed. The new orders will
be placed atomically in that they will all be placed or none will be placed.
In the case where the new orders cannot be placed the cancellations will not
be rolled back. See ReplaceInstruction.
}

