% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LimitOnCloseOrder}
\alias{LimitOnCloseOrder}
\title{LimitOnCloseOrder}
\usage{
LimitOnCloseOrder(liability, price)
}
\arguments{
\item{liability}{is mandatory of type numeric.
The size of the bet.}

\item{price}{is mandatory of type numeric.
The limit price of the bet if LOC}
}
\description{
Place a new LIMIT_ON_CLOSE bet
}

