% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LimitOrder}
\alias{LimitOrder}
\title{LimitOrder}
\usage{
LimitOrder(size, price, persistenceType)
}
\arguments{
\item{size}{is mandatory of type numeric.
The size of the bet.}

\item{price}{is mandatory of type numeric.
The limit price}

\item{persistenceType}{is mandatory of type PersistenceType.
What to do with the order at turn-in-play}
}
\description{
Place a new LIMIT order (simple exchange bet for immediate execution)
}

