% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listAccountSubscriptionTokens}
\alias{listAccountSubscriptionTokens}
\title{listAccountSubscriptionTokens}
\usage{
listAccountSubscriptionTokens(exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\value{
List of subscription tokens associated with the account
\code{\link{AccountSubscription}}
}
\description{
List of subscription tokens associated with the account
}

