% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listApplicationSubscriptionTokens}
\alias{listApplicationSubscriptionTokens}
\title{listApplicationSubscriptionTokens}
\usage{
listApplicationSubscriptionTokens(subscriptionStatus = NULL,
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{subscriptionStatus}{is optional of type SubscriptionStatus.
Optionally filter response by Subscription status of the token}
}
\value{
List of subscription tokens for an application
\code{\link{ApplicationSubscription}}
}
\description{
List of subscription tokens for an application
}

