% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listCurrencyRates}
\alias{listCurrencyRates}
\title{listCurrencyRates}
\usage{
listCurrencyRates(fromCurrency = NULL, exchange = .BetfairEnv$exchange,
  appKey = .BetfairEnv$appKey, sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{fromCurrency}{is optional of type character.
The currency from which the rates are computed. GBP is used if no
value is passed}
}
\value{
List of currency rates \code{\link{CurrencyRate}}
}
\description{
Returns a list of currency rates based on given currency
}

