% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listMarketCatalogue}
\alias{listMarketCatalogue}
\title{listMarketCatalogue}
\usage{
listMarketCatalogue(filter = MarketFilter(), marketProjection = NULL,
  sort = NULL, maxResults = i32(), locale = NULL,
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{filter}{is mandatory of type MarketFilter.
The filter to select desired markets. All markets that match the
criteria in the filter are selected.}

\item{marketProjection}{is optional of type SetOfMarketProjection.
The type and amount of data returned about the market.}

\item{sort}{is optional of type MarketSort.
The order of the results. Will default to RANK if not passed}

\item{maxResults}{is mandatory of type integer.
limit on the total number of results returned, must be greater than 0
and less than or equal to 1000}

\item{locale}{is optional of type character.
The language used for the response. If not specified, the default is
returned.}
}
\value{
output data.frame
}
\description{
Returns a list of information about markets that does not change (or changes
very rarely). You use listMarketCatalogue to retrieve the name of the market,
the names of selections and other information about markets.
}
\examples{
# Example is for horse racing using pure R interface
eventTypes<-r.listEventTypes(filter=MarketFilter(textQuery="Horse Racing"))
horseRacingTypeId<-eventTypes$result[[1]]$eventType$id

# Get the description of the first market using the catalogue
# Uses R and types checking
firstRace<-r.listMarketCatalogue(
 filter=MarketFilter(eventTypeIds=SetOfEventTypeId(horseRacingTypeId),
                     marketCountries=SetOfCountryCode(countryCode),
                     marketTypeCodes=SetOfMarketType("WIN"),
                     marketStartTime=TimeRange(from=format(Sys.Date(),"\%Y-\%m-\%dT\%H:\%M:\%SZ"))),
 marketProjection=SetOfMarketProjection('RUNNER_DESCRIPTION','MARKET_START_TIME'),
 sort=MarketSort("FIRST_TO_START"),
 maxResults=1
)

# Get first 10 todays uk horse racing WIN markets
winHorseRacing<-listMarketCatalogue(
  filter=list(eventTypeIds=list(horseRacingTypeId),
              marketCountries=list(countryCode),
              marketTypeCodes=list("WIN"),
              marketStartTime=list(from=format(Sys.time(),"\%Y-\%m-\%dT\%H:\%M:\%SZ",tz='GMT'))),
  marketProjection=list('EVENT','RUNNER_DESCRIPTION', 'MARKET_START_TIME'),
  sort="LAST_TO_START",
  maxResults=10
)
}

