% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listMarketTypes}
\alias{listMarketTypes}
\title{listMarketTypes}
\usage{
listMarketTypes(filter = MarketFilter(), locale = NULL,
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{filter}{is mandatory of type MarketFilter.
The filter to select desired markets. All markets that match the
criteria in the filter are selected.}

\item{locale}{is optional of type character.
The language used for the response. If not specified, the default is
returned.}
}
\value{
output data
}
\description{
Returns a list of market types (i.e. MATCH_ODDS, NEXT_GOAL) associated with
the markets selected by the MarketFilter. The market types are always the
same, regardless of locale.
}
\examples{
listMarketTypes(filter=MarketFilter(textQuery="Horse Racing"))
}

