% Generated by roxygen2 (4.0.2): do not edit by hand
\name{listTimeRanges}
\alias{listTimeRanges}
\title{listTimeRanges}
\usage{
listTimeRanges(filter = MarketFilter(), granularity = TimeGranularity(),
  exchange = .BetfairEnv$exchange, appKey = .BetfairEnv$appKey,
  sessionToken = .BetfairEnv$sessionToken)
}
\arguments{
\item{filter}{is mandatory of type MarketFilter.
The filter to select desired markets. All markets that match the
criteria in the filter are selected.}

\item{granularity}{is mandatory of type TimeGranularity.
The granularity of time periods that correspond to markets selected by
the market filter.}
}
\value{
output data.frame
}
\description{
Returns a list of time ranges in the granularity specified in the request
(i.e. 3PM to 4PM, Aug 14th to Aug 15th) associated with the markets selected
by the MarketFilter.
}
\note{
Even though the xml specifies granularity optional you need to include it
}
\examples{
# Include granularity
listTimeRanges(granularity=TimeGranularity('DAYS'))
}

