% Generated by roxygen2 (4.0.2): do not edit by hand
\name{login}
\alias{login}
\title{login}
\usage{
login(appKey,username,password)
}
\arguments{
\item{appKey}{product or application key}

\item{username}{betfair username}

\item{password}{betfair password}
}
\value{
list(status,token,product,status,error)
}
\description{
Login with no certificate
}

