% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MarketBettingTypeValidValues}
\alias{MarketBettingTypeValidValues}
\title{MarketBettingType}
\format{\preformatted{ chr [1:6] "ODDS" "LINE" "RANGE" "ASIAN_HANDICAP_DOUBLE_LINE" "ASIAN_HANDICAP_SINGLE_LINE" "FIXED_ODDS"
}}
\usage{
MarketBettingType(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
MarketBettingType
}
\details{
{
  \itemize{
    \item {     ODDS
           Odds Market
    }
    \item {     LINE
           Line Market
    }
    \item {     RANGE
           Range Market
    }
    \item {     ASIAN_HANDICAP_DOUBLE_LINE
           Asian Handicap Market
    }
    \item {     ASIAN_HANDICAP_SINGLE_LINE
           Asian Single Line Market
    }
    \item {     FIXED_ODDS
           Sportsbook Odds Market. This type is deprecated and will be removed in
           future releases, when Sportsbook markets will be represented as ODDS
           market but with a different product type.
    }
  }
}
}
\keyword{datasets}

