% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketDescription}
\alias{MarketDescription}
\title{MarketDescription}
\usage{
MarketDescription(persistenceEnabled, bspMarket, marketTime, suspendTime,
  settleTime = NULL, bettingType, turnInPlayEnabled, marketType, regulator,
  marketBaseRate, discountAllowed, wallet = NULL, rules = NULL,
  rulesHasDate = NULL, clarifications = NULL)
}
\arguments{
\item{persistenceEnabled}{is mandatory of type logical.
persistenceMarket}

\item{bspMarket}{is mandatory of type logical.
bspMarket}

\item{marketTime}{is mandatory of type dateTime.
marketTime}

\item{suspendTime}{is mandatory of type dateTime.
marketSuspendTime}

\item{settleTime}{is optional of type dateTime.
settled time}

\item{bettingType}{is mandatory of type character.
bettingType}

\item{turnInPlayEnabled}{is mandatory of type logical.
canTurnInPlay}

\item{marketType}{is mandatory of type character.
market base type}

\item{regulator}{is mandatory of type character.
the market regulator}

\item{marketBaseRate}{is mandatory of type numeric.}

\item{discountAllowed}{is mandatory of type logical.}

\item{wallet}{is optional of type character.}

\item{rules}{is optional of type character.}

\item{rulesHasDate}{is optional of type logical.}

\item{clarifications}{is optional of type character.}
}
\description{
Market definition
}

