% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketFilter}
\alias{MarketFilter}
\title{MarketFilter}
\usage{
MarketFilter(textQuery = NULL, exchangeIds = NULL, eventTypeIds = NULL,
  eventIds = NULL, competitionIds = NULL, marketIds = NULL,
  venues = NULL, bspOnly = NULL, turnInPlayEnabled = NULL,
  inPlayOnly = NULL, marketBettingTypes = NULL, marketCountries = NULL,
  marketTypeCodes = NULL, marketStartTime = NULL, withOrders = NULL)
}
\arguments{
\item{textQuery}{is optional of type character.
Restrict markets by any text associated with the market such as the
Name, Event, Competition, etc. You can include a wildcard (*)
character as long as it is not the first character.}

\item{exchangeIds}{is optional of type SetOfExchangeId.
Restrict markets by the Exchange where the market operates. Note: This
is not currently in use and only entities for the current exchange
will be returned.}

\item{eventTypeIds}{is optional of type SetOfEventTypeId.
Restrict markets by event type associated with the market. (i.e.,
Football, Hockey, etc)}

\item{eventIds}{is optional of type SetOfEventId.
Restrict markets by the event id associated with the market.}

\item{competitionIds}{is optional of type SetOfCompetitionId.
Restrict markets by the competitions associated with the market.}

\item{marketIds}{is optional of type SetOfMarketId.
Restrict markets by the market id associated with the market.}

\item{venues}{is optional of type SetOfVenue.
Restrict markets by the venue associated with the market. Currently
only Horse Racing markets have venues.}

\item{bspOnly}{is optional of type logical.
Restrict to bsp markets only, if True or non-bsp markets if False. If
not specified then returns both BSP and non-BSP markets}

\item{turnInPlayEnabled}{is optional of type logical.
Restrict to markets that will turn in play if True or will not turn in
play if false. If not specified, returns both.}

\item{inPlayOnly}{is optional of type logical.
Restrict to markets that will turn in play if True or will not turn in
play if false. If not specified, returns both.}

\item{marketBettingTypes}{is optional of type SetOfMarketBettingType.
Restrict to markets that match the betting type of the market (i.e.
Odds, Asian Handicap Singles, or Asian Handicap Doubles}

\item{marketCountries}{is optional of type SetOfCountryCode.
Restrict to markets that are in the specified country or countries}

\item{marketTypeCodes}{is optional of type SetOfMarketType.
Restrict to markets that match the type of the market (i.e.,
MATCH_ODDS, HALF_TIME_SCORE). You should use this instead of relying
on the market name as the market type codes are the same in all
locales}

\item{marketStartTime}{is optional of type TimeRange.
Restrict to markets with a market start time before or after the
specified date}

\item{withOrders}{is optional of type SetOfOrderStatus.
Restrict to markets that I have one or more orders in these status.}
}
\description{
MarketFilter
}

