% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketLicence}
\alias{MarketLicence}
\title{MarketLicence}
\usage{
MarketLicence(wallet, rules = NULL, rulesHasDate = NULL,
  clarifications = NULL)
}
\arguments{
\item{wallet}{is mandatory of type character.
The wallet from which funds will be taken when betting on this market}

\item{rules}{is optional of type character.
The rules for this market}

\item{rulesHasDate}{is optional of type logical.
The market's start date and time are relevant to the rules.}

\item{clarifications}{is optional of type character.
Clarifications to the rules for the market}
}
\description{
Market Licence
}

