% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketLineRangeInfo}
\alias{MarketLineRangeInfo}
\title{MarketLineRangeInfo}
\usage{
MarketLineRangeInfo(maxUnitValue, minUnitValue, interval, marketUnit)
}
\arguments{
\item{maxUnitValue}{is mandatory of type numeric.
maxPrice}

\item{minUnitValue}{is mandatory of type numeric.
minPrice}

\item{interval}{is mandatory of type numeric.
interval}

\item{marketUnit}{is mandatory of type character.
unit}
}
\description{
Market Line and Range Info
}

