% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketOnCloseOrder}
\alias{MarketOnCloseOrder}
\title{MarketOnCloseOrder}
\usage{
MarketOnCloseOrder(liability)
}
\arguments{
\item{liability}{is mandatory of type numeric.
The size of the bet.}
}
\description{
Place a new MARKET_ON_CLOSE bet
}

