% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketProfitAndLoss}
\alias{MarketProfitAndLoss}
\title{MarketProfitAndLoss}
\usage{
MarketProfitAndLoss(marketId = NULL, commissionApplied = NULL,
  profitAndLosses = NULL)
}
\arguments{
\item{marketId}{is optional of type character.
The unique identifier for the market}

\item{commissionApplied}{is optional of type numeric.
The commission rate applied to P&L values. Only returned if
netOfCommision option is requested}

\item{profitAndLosses}{is optional of type ListOfRunnerProfitAndLoss.
Calculated profit and loss data.}
}
\description{
Profit and loss in a market
}

