% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MarketProjectionValidValues}
\alias{MarketProjectionValidValues}
\title{MarketProjection}
\format{\preformatted{ chr [1:7] "COMPETITION" "EVENT" "EVENT_TYPE" "MARKET_START_TIME" "MARKET_DESCRIPTION" "RUNNER_DESCRIPTION" ...
}}
\usage{
MarketProjection(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
MarketProjection
}
\details{
{
  \itemize{
    \item {     COMPETITION
           If not selected then the competition will not be returned with
           marketCatalogue
    }
    \item {     EVENT
           If not selected then the event will not be returned with
           marketCatalogue
    }
    \item {     EVENT_TYPE
           If not selected then the eventType will not be returned with
           marketCatalogue
    }
    \item {     MARKET_START_TIME
           If not selected then the start time will not be returned with
           marketCatalogue
    }
    \item {     MARKET_DESCRIPTION
           If not selected then the description will not be returned with
           marketCatalogue
    }
    \item {     RUNNER_DESCRIPTION
           If not selected then the runners will not be returned with
           marketCatalogue
    }
    \item {     RUNNER_METADATA
           If not selected then the runner metadata will not be returned with
           marketCatalogue. If selected then RUNNER_DESCRIPTION will also be
           returned regardless of whether it is included as a market projection.
    }
  }
}
}
\keyword{datasets}

