% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketRates}
\alias{MarketRates}
\title{MarketRates}
\usage{
MarketRates(marketBaseRate, discountAllowed)
}
\arguments{
\item{marketBaseRate}{is mandatory of type numeric.
marketBaseRate}

\item{discountAllowed}{is mandatory of type logical.
discountAllowed}
}
\description{
Market Rates
}

