% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MarketSortValidValues}
\alias{MarketSortValidValues}
\title{MarketSort}
\format{\preformatted{ chr [1:6] "MINIMUM_TRADED" "MAXIMUM_TRADED" "MINIMUM_AVAILABLE" "MAXIMUM_AVAILABLE" "FIRST_TO_START" ...
}}
\usage{
MarketSort(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
MarketSort
}
\details{
{
  \itemize{
    \item {     MINIMUM_TRADED
           Minimum traded volume
    }
    \item {     MAXIMUM_TRADED
           Maximum traded volume
    }
    \item {     MINIMUM_AVAILABLE
           Minimum available to match
    }
    \item {     MAXIMUM_AVAILABLE
           Maximum available to match
    }
    \item {     FIRST_TO_START
           The closest markets based on their expected start time
    }
    \item {     LAST_TO_START
           The most distant markets based on their expected start time
    }
  }
}
}
\keyword{datasets}

