% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketState}
\alias{MarketState}
\title{MarketState}
\usage{
MarketState(status, betDelay, bspReconciled, complete, inplay,
  numberOfActiveRunners, lastMatchTime, totalMatched, totalAvailable)
}
\arguments{
\item{status}{is mandatory of type character.
marketStatus}

\item{betDelay}{is mandatory of type integer.
betDelay}

\item{bspReconciled}{is mandatory of type logical.
bspReconciled}

\item{complete}{is mandatory of type logical.
complete}

\item{inplay}{is mandatory of type logical.
inplay}

\item{numberOfActiveRunners}{is mandatory of type integer.
numberActiveOfRunners}

\item{lastMatchTime}{is mandatory of type dateTime.
lastMatchTime}

\item{totalMatched}{is mandatory of type numeric.}

\item{totalAvailable}{is mandatory of type numeric.
Zero for closed markets}
}
\description{
Market definition
}

