% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MarketStatusValidValues}
\alias{MarketStatusValidValues}
\title{MarketStatus}
\format{\preformatted{ chr [1:4] "INACTIVE" "OPEN" "SUSPENDED" "CLOSED"
}}
\usage{
MarketStatus(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
MarketStatus
}
\details{
{
  \itemize{
    \item {     INACTIVE
           Inactive Market
    }
    \item {     OPEN
           Open Market
    }
    \item {     SUSPENDED
           Suspended Market
    }
    \item {     CLOSED
           Closed Market
    }
  }
}
}
\keyword{datasets}

