% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketTypeResult}
\alias{MarketTypeResult}
\title{MarketTypeResult}
\usage{
MarketTypeResult(marketType = NULL, marketCount = NULL)
}
\arguments{
\item{marketType}{is optional of type character.
Market Type}

\item{marketCount}{is optional of type integer.
Count of markets associated with this marketType}
}
\description{
MarketType Result
}

