% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MarketVersion}
\alias{MarketVersion}
\title{MarketVersion}
\usage{
MarketVersion(version = NULL)
}
\arguments{
\item{version}{is optional of type integer.
A non-monotonically increasing number indicating market changes}
}
\description{
Market version
}

