% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Match}
\alias{Match}
\title{Match}
\usage{
Match(betId = NULL, matchId = NULL, side, price, Size, matchDate = NULL)
}
\arguments{
\item{betId}{is optional of type character.
Only present if no rollup}

\item{matchId}{is optional of type character.
Only present if no rollup}

\item{side}{is mandatory of type character.}

\item{price}{is mandatory of type numeric.
Either actual match price or avg match price depending on rollup.}

\item{Size}{is mandatory of type numeric.
Size matched at in this fragment, or at this price or avg price
depending on rollup}

\item{matchDate}{is optional of type dateTime.
Only present if no rollup}
}
\description{
An individual bet Match, or rollup by price or avg price. Rollup depends on
the requested MatchProjection
}

