% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MatchProjectionValidValues}
\alias{MatchProjectionValidValues}
\title{MatchProjection}
\format{\preformatted{ chr [1:3] "NO_ROLLUP" "ROLLED_UP_BY_PRICE" "ROLLED_UP_BY_AVG_PRICE"
}}
\usage{
MatchProjection(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
MatchProjection
}
\details{
{
  \itemize{
    \item {     NO_ROLLUP
    }
    \item {     ROLLED_UP_BY_PRICE
    }
    \item {     ROLLED_UP_BY_AVG_PRICE
    }
  }
}
}
\keyword{datasets}

