% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Order}
\alias{Order}
\title{Order}
\usage{
Order(betId, orderType, status, persistenceType, side, price, size,
  bspLiability, placedDate, avgPriceMatched = NULL, sizeMatched = NULL,
  sizeRemaining = NULL, sizeLapsed = NULL, sizeCancelled = NULL,
  sizeVoided = NULL)
}
\arguments{
\item{betId}{is mandatory of type character.}

\item{orderType}{is mandatory of type character.
BSP Order type.}

\item{status}{is mandatory of type character.
Either EXECUTABLE (an unmatched amount remains) or EXECUTION_COMPLETE
(no unmatched amount remains).}

\item{persistenceType}{is mandatory of type character.
What to do with the order at turn-in-play}

\item{side}{is mandatory of type character.}

\item{price}{is mandatory of type numeric.
The price of the bet.}

\item{size}{is mandatory of type numeric.
The size of the bet.}

\item{bspLiability}{is mandatory of type numeric.
Not to be confused with size. This is the liability of a given BSP
bet.}

\item{placedDate}{is mandatory of type dateTime.
The date, to the second, the bet was placed.}

\item{avgPriceMatched}{is optional of type numeric.
The average price matched at. Voided match fragments are removed from
this average calculation.}

\item{sizeMatched}{is optional of type numeric.
The current amount of this bet that was matched.}

\item{sizeRemaining}{is optional of type numeric.
The current amount of this bet that is unmatched.}

\item{sizeLapsed}{is optional of type numeric.
The current amount of this bet that was lapsed.}

\item{sizeCancelled}{is optional of type numeric.
The current amount of this bet that was cancelled.}

\item{sizeVoided}{is optional of type numeric.
The current amount of this bet that was voided.}
}
\description{
Order
}

