% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{OrderByValidValues}
\alias{OrderByValidValues}
\title{OrderBy}
\format{\preformatted{ chr [1:6] "BY_BET" "BY_MARKET" "BY_PLACE_TIME" "BY_MATCH_TIME" "BY_VOID_TIME" "BY_SETTLED_TIME"
}}
\usage{
OrderBy(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
OrderBy
}
\details{
{
  \itemize{
    \item {     BY_BET
           Deprecated Use BY_PLACE_TIME instead. Order by placed time, then bet
           id.
    }
    \item {     BY_MARKET
           Order by market id, then placed time, then bet id.
    }
    \item {     BY_PLACE_TIME
           Order by placed time, then bet id. This is an alias of to be
           deprecated BY_BET.
    }
    \item {     BY_MATCH_TIME
           Order by time of last matched fragment (if any), then placed time,
           then bet id. Filters out orders which have no matched date
    }
    \item {     BY_VOID_TIME
           Order by time of last voided fragment (if any), then by last match
           time, then placed time, then bet id. Filters out orders which have not
           been voided.
    }
    \item {     BY_SETTLED_TIME
           Order by time of last settled fragment (if any), then by last match
           time, then placed time, then bet id. Filters out orders which have not
           been settled.
    }
  }
}
}
\keyword{datasets}

