% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{OrderProjectionValidValues}
\alias{OrderProjectionValidValues}
\title{OrderProjection}
\format{\preformatted{ chr [1:3] "ALL" "EXECUTABLE" "EXECUTION_COMPLETE"
}}
\usage{
OrderProjection(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
OrderProjection
}
\details{
{
  \itemize{
    \item {     ALL
    }
    \item {     EXECUTABLE
    }
    \item {     EXECUTION_COMPLETE
    }
  }
}
}
\keyword{datasets}

