% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{OrderStatusValidValues}
\alias{OrderStatusValidValues}
\title{OrderStatus}
\format{\preformatted{ chr [1:2] "EXECUTION_COMPLETE" "EXECUTABLE"
}}
\usage{
OrderStatus(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
OrderStatus
}
\details{
{
  \itemize{
    \item {     EXECUTION_COMPLETE
           An order that does not have any remaining unmatched portion.
    }
    \item {     EXECUTABLE
           An order that has a remaining unmatched portion.
    }
  }
}
}
\keyword{datasets}

