% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{OrderTypeValidValues}
\alias{OrderTypeValidValues}
\title{OrderType}
\format{\preformatted{ chr [1:3] "LIMIT" "LIMIT_ON_CLOSE" "MARKET_ON_CLOSE"
}}
\usage{
OrderType(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
OrderType
}
\details{
{
  \itemize{
    \item {     LIMIT
           A normal exchange limit order for immediate execution
    }
    \item {     LIMIT_ON_CLOSE
           Limit order for the auction (SP)
    }
    \item {     MARKET_ON_CLOSE
           Market order for the auction (SP)
    }
  }
}
}
\keyword{datasets}

