% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{PersistenceTypeValidValues}
\alias{PersistenceTypeValidValues}
\title{PersistenceType}
\format{\preformatted{ chr [1:3] "LAPSE" "PERSIST" "MARKET_ON_CLOSE"
}}
\usage{
PersistenceType(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
PersistenceType
}
\details{
{
  \itemize{
    \item {     LAPSE
           Lapse the order at turn-in-play
    }
    \item {     PERSIST
           Persist the order to in-play
    }
    \item {     MARKET_ON_CLOSE
           Put the order into the auction (SP) at turn-in-play
    }
  }
}
}
\keyword{datasets}

