% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlaceInstruction}
\alias{PlaceInstruction}
\title{PlaceInstruction}
\usage{
PlaceInstruction(orderType, selectionId, handicap = NULL, side,
  limitOrder = NULL, limitOnCloseOrder = NULL, marketOnCloseOrder = NULL)
}
\arguments{
\item{orderType}{is mandatory of type OrderType.}

\item{selectionId}{is mandatory of type integer.
The selection_id.}

\item{handicap}{is optional of type numeric.
The handicap applied to the selection, if on an asian-style market.}

\item{side}{is mandatory of type Side.
Back or Lay}

\item{limitOrder}{is optional of type LimitOrder.}

\item{limitOnCloseOrder}{is optional of type LimitOnCloseOrder.}

\item{marketOnCloseOrder}{is optional of type MarketOnCloseOrder.}
}
\description{
Instruction to place a new order
}

