% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlaceInstructionReport}
\alias{PlaceInstructionReport}
\title{PlaceInstructionReport}
\usage{
PlaceInstructionReport(status, errorCode = NULL, instruction, betId = NULL,
  placedDate = NULL, averagePriceMatched = NULL, sizeMatched = NULL)
}
\arguments{
\item{status}{is mandatory of type character.
whether the command succeeded or failed}

\item{errorCode}{is optional of type character.
cause of failure, or null if command succeeds}

\item{instruction}{is mandatory of type PlaceInstruction.
The instruction that was requested}

\item{betId}{is optional of type character.
The bet ID of the new bet. May be null on failure.}

\item{placedDate}{is optional of type dateTime.}

\item{averagePriceMatched}{is optional of type numeric.}

\item{sizeMatched}{is optional of type numeric.}
}
\description{
Response to a PlaceInstruction
}

