% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{PriceDataValidValues}
\alias{PriceDataValidValues}
\title{PriceData}
\format{\preformatted{ chr [1:5] "SP_AVAILABLE" "SP_TRADED" "EX_BEST_OFFERS" "EX_ALL_OFFERS" "EX_TRADED"
}}
\usage{
PriceData(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
PriceData
}
\details{
{
  \itemize{
    \item {     SP_AVAILABLE
    }
    \item {     SP_TRADED
    }
    \item {     EX_BEST_OFFERS
    }
    \item {     EX_ALL_OFFERS
           EX_ALL_OFFERS trumps EX_BEST_OFFERS if both settings are present
    }
    \item {     EX_TRADED
    }
  }
}
}
\keyword{datasets}

