% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PriceProjection}
\alias{PriceProjection}
\title{PriceProjection}
\usage{
PriceProjection(priceData = NULL, exBestOffersOverrides = NULL,
  virtualise = NULL, rolloverStakes = NULL)
}
\arguments{
\item{priceData}{is optional of type SetOfPriceData.
The basic price data you want to receive in the response.}

\item{exBestOffersOverrides}{is optional of type ExBestOffersOverrides.
Options to alter the default representation of best offer prices
Applicable to EX_BEST_OFFERS priceData selection}

\item{virtualise}{is optional of type logical.
Indicates if the returned prices should include virtual prices.
Applicable to EX_BEST_OFFERS and EX_ALL_OFFERS priceData selections,
default value is false.}

\item{rolloverStakes}{is optional of type logical.
Indicates if the volume returned at each price point should be the
absolute value or a cumulative sum of volumes available at the price
and all better prices. If unspecified defaults to false. Applicable to
EX_BEST_OFFERS and EX_ALL_OFFERS price projections. Not supported as
yet.}
}
\description{
Selection criteria of the returning price data
}

