% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReplaceExecutionReport}
\alias{ReplaceExecutionReport}
\title{ReplaceExecutionReport}
\usage{
ReplaceExecutionReport(customerRef = NULL, status, errorCode = NULL,
  marketId = NULL, instructionReports = NULL)
}
\arguments{
\item{customerRef}{is optional of type character.
Echo of the customerRef if passed.}

\item{status}{is mandatory of type character.}

\item{errorCode}{is optional of type character.}

\item{marketId}{is optional of type character.
Echo of marketId passed}

\item{instructionReports}{is optional of type ListOfReplaceInstructionReport.}
}
\description{
ReplaceExecutionReport
}

