% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReplaceInstruction}
\alias{ReplaceInstruction}
\title{ReplaceInstruction}
\usage{
ReplaceInstruction(betId, newPrice)
}
\arguments{
\item{betId}{is mandatory of type character.}

\item{newPrice}{is mandatory of type numeric.
The price to replace the bet at}
}
\description{
Instruction to replace a LIMIT or LIMIT_ON_CLOSE order at a new price.
Original order will be cancelled and a new order placed at the new price for
the remaining stake.
}

