% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReplaceInstructionReport}
\alias{ReplaceInstructionReport}
\title{ReplaceInstructionReport}
\usage{
ReplaceInstructionReport(status, errorCode = NULL,
  cancelInstructionReport = NULL, placeInstructionReport = NULL)
}
\arguments{
\item{status}{is mandatory of type character.
whether the command succeeded or failed}

\item{errorCode}{is optional of type character.
cause of failure, or null if command succeeds}

\item{cancelInstructionReport}{is optional of type CancelInstructionReport.
Cancelation report for the original order}

\item{placeInstructionReport}{is optional of type PlaceInstructionReport.
Placement report for the new order}
}
\description{
ReplaceInstructionReport
}

