% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{RollupModelValidValues}
\alias{RollupModelValidValues}
\title{RollupModel}
\format{\preformatted{ chr [1:4] "STAKE" "PAYOUT" "MANAGED_LIABILITY" "NONE"
}}
\usage{
RollupModel(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
RollupModel
}
\details{
{
  \itemize{
    \item {     STAKE
           The volumes will be rolled up to the minimum value which is >=
           rollupLimit.
    }
    \item {     PAYOUT
           The volumes will be rolled up to the minimum value where the payout(
           price * volume ) is >= rollupLimit.
    }
    \item {     MANAGED_LIABILITY
           The volumes will be rolled up to the minimum value which is >=
           rollupLimit, until a lay price threshold. There after, the volumes
           will be rolled up to the minimum value such that the liability >= a
           minimum liability. Not supported as yet.
    }
    \item {     NONE
           No rollup will be applied. However the volumes will be filtered by
           currency specific minimum stake unless overridden specifically for the
           channel.
    }
  }
}
}
\keyword{datasets}

