% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Runner}
\alias{Runner}
\title{Runner}
\usage{
Runner(selectionId, handicap, status, adjustmentFactor,
  lastPriceTraded = NULL, totalMatched = NULL, removalDate = NULL,
  sp = NULL, ex = NULL, orders = NULL, matches = NULL)
}
\arguments{
\item{selectionId}{is mandatory of type integer.
The unique id of the runner (selection)}

\item{handicap}{is mandatory of type numeric.
The handicap}

\item{status}{is mandatory of type character.
The status of the selection (i.e., ACTIVE, REMOVED, WINNER, LOSER)}

\item{adjustmentFactor}{is mandatory of type numeric.
The adjustment factor applied if the selection is removed}

\item{lastPriceTraded}{is optional of type numeric.
The price of the most recent bet matched on this selection}

\item{totalMatched}{is optional of type numeric.
The total amount matched on this runner. This value is truncated at
2dp.}

\item{removalDate}{is optional of type dateTime.
If date and time the runner was removed}

\item{sp}{is optional of type StartingPrices.
The BSP related prices for this runner}

\item{ex}{is optional of type ExchangePrices.
The Exchange prices available for this runner}

\item{orders}{is optional of type ListOfOrder.
List of orders in the market}

\item{matches}{is optional of type ListOfMatch.
List of matches (i.e, orders that have been fully or partially
executed)}
}
\description{
The dynamic data about runners in a market
}

