% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RunnerCatalog}
\alias{RunnerCatalog}
\title{RunnerCatalog}
\usage{
RunnerCatalog(selectionId, runnerName, handicap, sortPriority,
  metadata = NULL)
}
\arguments{
\item{selectionId}{is mandatory of type integer.
The unique id for the selection.}

\item{runnerName}{is mandatory of type character.
The name of the runner}

\item{handicap}{is mandatory of type numeric.
The handicap}

\item{sortPriority}{is mandatory of type integer.
The sort priority of this runner}

\item{metadata}{is optional of type map(string,string).
Metadata associated with the runner}
}
\description{
Information about the Runners (selections) in a market
}

