% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RunnerDescription}
\alias{RunnerDescription}
\title{RunnerDescription}
\usage{
RunnerDescription(runnerName, metadata = NULL)
}
\arguments{
\item{runnerName}{is mandatory of type character.
The name of the runner}

\item{metadata}{is optional of type map(string,string).
The metadata for this runner, such as horse racing silks.}
}
\description{
RunnerDescription
}

