% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RunnerId}
\alias{RunnerId}
\title{RunnerId}
\usage{
RunnerId(marketId, selectionId, handicap = NULL)
}
\arguments{
\item{marketId}{is mandatory of type character.
The id of the market bet on}

\item{selectionId}{is mandatory of type integer.
The id of the selection bet on}

\item{handicap}{is optional of type numeric.
The handicap associated with the runner in case of asian handicap
markets, null otherwise.}
}
\description{
This object contains the unique identifier for a runner
}

