% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RunnerProfitAndLoss}
\alias{RunnerProfitAndLoss}
\title{RunnerProfitAndLoss}
\usage{
RunnerProfitAndLoss(selectionId = NULL, ifWin = NULL, ifLose = NULL)
}
\arguments{
\item{selectionId}{is optional of type integer.
The unique identifier for the selection}

\item{ifWin}{is optional of type numeric.
Profit and loss for the market if this selection is the winner}

\item{ifLose}{is optional of type numeric.
Profit and loss for the market if this selection is the loser. Only
returned for multi-winner odds markets.}
}
\description{
Profit and loss if selection is wins or loses
}

