% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{RunnerStatusValidValues}
\alias{RunnerStatusValidValues}
\title{RunnerStatus}
\format{\preformatted{ chr [1:5] "ACTIVE" "WINNER" "LOSER" "REMOVED_VACANT" "REMOVED"
}}
\usage{
RunnerStatus(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
RunnerStatus
}
\details{
{
  \itemize{
    \item {     ACTIVE
           ACTIVE
    }
    \item {     WINNER
           WINNER
    }
    \item {     LOSER
           LOSER
    }
    \item {     REMOVED_VACANT
    }
    \item {     REMOVED
           REMOVED
    }
  }
}
}
\keyword{datasets}

