% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SideValidValues}
\alias{SideValidValues}
\title{Side}
\format{\preformatted{ chr [1:2] "BACK" "LAY"
}}
\usage{
Side(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
Side
}
\details{
{
  \itemize{
    \item {     BACK
           To back a team, horse or outcome is to bet on the selection to win.
    }
    \item {     LAY
           To lay a team, horse, or outcome is to bet on the selection to lose.
    }
  }
}
}
\keyword{datasets}

