% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SortDirValidValues}
\alias{SortDirValidValues}
\title{SortDir}
\format{\preformatted{ chr [1:2] "EARLIEST_TO_LATEST" "LATEST_TO_EARLIEST"
}}
\usage{
SortDir(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
SortDir
}
\details{
{
  \itemize{
    \item {     EARLIEST_TO_LATEST
           Order from earliest value to latest e.g. lowest betId is first in the
           results.
    }
    \item {     LATEST_TO_EARLIEST
           Order from the latest value to the earliest e.g. highest betId is
           first in the results.
    }
  }
}
}
\keyword{datasets}

