% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ssologin}
\alias{ssologin}
\title{login}
\usage{
login(appKey,username,password,sslcert)
}
\arguments{
\item{appKey}{product or application key}

\item{username}{betfair username}

\item{password}{betfair password}

\item{sslcert}{}
}
\value{
list(status,token,product,status,error)
}
\description{
Login with certificate
}

