% Generated by roxygen2 (4.0.2): do not edit by hand
\name{StartingPrices}
\alias{StartingPrices}
\title{StartingPrices}
\usage{
StartingPrices(nearPrice = NULL, farPrice = NULL, backStakeTaken = NULL,
  layLiabilityTaken = NULL, actualSP = NULL)
}
\arguments{
\item{nearPrice}{is optional of type numeric.
What the starting price would be if the market was reconciled now
taking into account the SP bets as well as unmatched exchange bets on
the same selection in the exchange.}

\item{farPrice}{is optional of type numeric.
What the starting price would be if the market was reconciled now
taking into account only the currently place SP bets. The Far Price is
not as complicated but not as accurate and only accounts for money on
the exchange at SP.}

\item{backStakeTaken}{is optional of type ListOfPriceSize.}

\item{layLiabilityTaken}{is optional of type ListOfPriceSize.}

\item{actualSP}{is optional of type numeric.
The final BSP price for this runner. Only available for a BSP market
that has been reconciled.}
}
\description{
Information about the Betfair Starting Price. Only available in BSP markets
}

