% Generated by roxygen2 (4.0.2): do not edit by hand
\name{StatementItem}
\alias{StatementItem}
\title{StatementItem}
\usage{
StatementItem(refId = NULL, itemDate, amount = NULL, balance = NULL,
  itemClass = NULL, itemClassData = NULL, legacyData = NULL)
}
\arguments{
\item{refId}{is optional of type character.
An external reference, eg. equivalent to betId in the case of an
exchange bet statement item.}

\item{itemDate}{is mandatory of type dateTime.
The date and time of the statement item, eg. equivalent to settledData
for an exchange bet statement item. (in ISO-8601 format, not
translated)}

\item{amount}{is optional of type numeric.
The amount of money the balance is adjusted by}

\item{balance}{is optional of type numeric.
Account balance.}

\item{itemClass}{is optional of type ItemClass.
Class of statement item. This value will determine which set of keys
will be included in itemClassData}

\item{itemClassData}{is optional of type map(string,string).
Key value pairs describing the current statement item. The set of keys
will be determined by the itemClass}

\item{legacyData}{is optional of type StatementLegacyData.
Set of fields originally returned from APIv6. Provided to facilitate
migration from APIv6 to API-NG, and ultimately onto itemClass and
itemClassData}
}
\description{
Summary of a cleared order.
}

