% Generated by roxygen2 (4.0.2): do not edit by hand
\name{StatementLegacyData}
\alias{StatementLegacyData}
\title{StatementLegacyData}
\usage{
StatementLegacyData(avgPrice = NULL, betSize = NULL, betType = NULL,
  betCategoryType = NULL, commissionRate = NULL, eventId = NULL,
  eventTypeId = NULL, fullMarketName = NULL, grossBetAmount = NULL,
  marketName = NULL, marketType = NULL, placedDate = NULL,
  selectionId = NULL, selectionName = NULL, startDate = NULL,
  transactionType = NULL, transactionId = NULL, winLose = NULL)
}
\arguments{
\item{avgPrice}{is optional of type numeric.}

\item{betSize}{is optional of type numeric.}

\item{betType}{is optional of type character.}

\item{betCategoryType}{is optional of type character.}

\item{commissionRate}{is optional of type character.}

\item{eventId}{is optional of type integer.}

\item{eventTypeId}{is optional of type integer.}

\item{fullMarketName}{is optional of type character.}

\item{grossBetAmount}{is optional of type numeric.}

\item{marketName}{is optional of type character.}

\item{marketType}{is optional of type character.}

\item{placedDate}{is optional of type dateTime.}

\item{selectionId}{is optional of type integer.}

\item{selectionName}{is optional of type character.}

\item{startDate}{is optional of type dateTime.}

\item{transactionType}{is optional of type character.}

\item{transactionId}{is optional of type integer.}

\item{winLose}{is optional of type character.}
}
\description{
Summary of a cleared order.
}

