% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{StatusValidValues}
\alias{StatusValidValues}
\title{Status}
\format{\preformatted{ chr "SUCCESS"
}}
\usage{
Status(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
Status
}
\details{
{
  \itemize{
    \item {     SUCCESS
           Sucess status
    }
  }
}
}
\keyword{datasets}

