% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SubscriptionHistory}
\alias{SubscriptionHistory}
\title{SubscriptionHistory}
\usage{
SubscriptionHistory(subscriptionToken, expiryDateTime = NULL,
  expiredDateTime = NULL, createdDateTime = NULL,
  activationDateTime = NULL, cancellationDateTime = NULL,
  subscriptionStatus = NULL, clientReference = NULL)
}
\arguments{
\item{subscriptionToken}{is mandatory of type character.
Application key identifier}

\item{expiryDateTime}{is optional of type dateTime.
Subscription Expiry date}

\item{expiredDateTime}{is optional of type dateTime.
Subscription Expired date}

\item{createdDateTime}{is optional of type dateTime.
Subscription Create date}

\item{activationDateTime}{is optional of type dateTime.
Subscription Activation date}

\item{cancellationDateTime}{is optional of type dateTime.
Subscription Cancelled date}

\item{subscriptionStatus}{is optional of type character.
Subscription status}

\item{clientReference}{is optional of type character.
Client reference}
}
\description{
Application subscription history details
}

