% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SubscriptionStatusValidValues}
\alias{SubscriptionStatusValidValues}
\title{SubscriptionStatus}
\format{\preformatted{ chr [1:5] "ALL" "ACTIVATED" "UNACTIVATED" "CANCELLED" "EXPIRED"
}}
\usage{
SubscriptionStatus(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
SubscriptionStatus
}
\details{
{
  \itemize{
    \item {     ALL
           Any subscription status
    }
    \item {     ACTIVATED
           Only activated subscriptions
    }
    \item {     UNACTIVATED
           Only unactivated subscriptions
    }
    \item {     CANCELLED
           Only cancelled subscriptions
    }
    \item {     EXPIRED
           Only expired subscriptions
    }
  }
}
}
\keyword{datasets}

