% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SubscriptionTokenInfo}
\alias{SubscriptionTokenInfo}
\title{SubscriptionTokenInfo}
\usage{
SubscriptionTokenInfo(subscriptionToken, activatedDateTime = NULL,
  expiryDateTime = NULL, expiredDateTime = NULL,
  cancellationDateTime = NULL, subscriptionStatus = NULL)
}
\arguments{
\item{subscriptionToken}{is mandatory of type character.
Subscription token}

\item{activatedDateTime}{is optional of type dateTime.
Subscription Activated date}

\item{expiryDateTime}{is optional of type dateTime.
Subscription Expiry date}

\item{expiredDateTime}{is optional of type dateTime.
Subscription Expired date}

\item{cancellationDateTime}{is optional of type dateTime.
Subscription Cancelled date}

\item{subscriptionStatus}{is optional of type character.
Subscription status}
}
\description{
Subscription token information
}

