% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TimeGranularityValidValues}
\alias{TimeGranularityValidValues}
\title{TimeGranularity}
\format{\preformatted{ chr [1:3] "DAYS" "HOURS" "MINUTES"
}}
\usage{
TimeGranularity(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
TimeGranularity
}
\details{
{
  \itemize{
    \item {     DAYS
    }
    \item {     HOURS
    }
    \item {     MINUTES
    }
  }
}
}
\keyword{datasets}

