% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TimeRange}
\alias{TimeRange}
\title{TimeRange}
\usage{
TimeRange(from = NULL, to = NULL)

TimeRange(from = NULL, to = NULL)
}
\arguments{
\item{from}{is optional of type dateTime.
from, format: ISO 8601)}

\item{to}{is optional of type dateTime.
to, format: ISO 8601}

\item{from}{is optional of type dateTime.
from, format: ISO 8601)}

\item{to}{is optional of type dateTime.
to, format: ISO 8601}
}
\description{
TimeRange

TimeRange
}

