% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TimeRangeResult}
\alias{TimeRangeResult}
\title{TimeRangeResult}
\usage{
TimeRangeResult(timeRange = NULL, marketCount = NULL)
}
\arguments{
\item{timeRange}{is optional of type TimeRange.
TimeRange}

\item{marketCount}{is optional of type integer.
Count of markets associated with this TimeRange}
}
\description{
TimeRange Result
}

