% Generated by roxygen2 (4.0.2): do not edit by hand
\name{UpdateInstruction}
\alias{UpdateInstruction}
\title{UpdateInstruction}
\usage{
UpdateInstruction(betId, newPersistenceType)
}
\arguments{
\item{betId}{is mandatory of type character.}

\item{newPersistenceType}{is mandatory of type PersistenceType.
The new persistence type to update this bet to}
}
\description{
Instruction to update LIMIT bet's persistence of an order that do not affect
exposure
}

