% Generated by roxygen2 (4.0.2): do not edit by hand
\name{UpdateInstructionReport}
\alias{UpdateInstructionReport}
\title{UpdateInstructionReport}
\usage{
UpdateInstructionReport(status, errorCode = NULL, instruction)
}
\arguments{
\item{status}{is mandatory of type character.
whether the command succeeded or failed}

\item{errorCode}{is optional of type character.
cause of failure, or null if command succeeds}

\item{instruction}{is mandatory of type UpdateInstruction.
The instruction that was requested}
}
\description{
UpdateInstructionReport
}

