% Generated by roxygen2 (4.0.2): do not edit by hand
\name{VenueResult}
\alias{VenueResult}
\title{VenueResult}
\usage{
VenueResult(venue = NULL, marketCount = NULL)
}
\arguments{
\item{venue}{is optional of type character.
Venue}

\item{marketCount}{is optional of type integer.
Count of markets associated with this Venue}
}
\description{
Venue Result
}

