% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WalletValidValues}
\alias{WalletValidValues}
\title{Wallet}
\format{\preformatted{ chr [1:2] "UK" "AUSTRALIAN"
}}
\usage{
Wallet(values)
}
\arguments{
\item{values}{Any of following valid values}
}
\description{
Wallet
}
\details{
{
  \itemize{
    \item {     UK
    }
    \item {     AUSTRALIAN
    }
  }
}
}
\keyword{datasets}

